Adding support mvbind
Note the current proposed change has the potential to add unintended consequences with helper packages. For instance, returning a list as lower_bound & upper_bound breaks the current implementation of pp_check_ordbeta.

I'm proposing the change in its current form to solicit feedback. An easier "simple" solution would be to return [0,1] or the bounds of the first response variable in the multivariate case and print a warning.